% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_models.R
\name{bernoulli_likelihood}
\alias{bernoulli_likelihood}
\title{Plot the likelihood function of a Bernoulli/binomial model}
\usage{
bernoulli_likelihood(n, m)
}
\arguments{
\item{n}{Number of trials}

\item{m}{Number of successes}
}
\value{
A ggplot object
}
\description{
Plot the likelihood function of a Bernoulli/binomial model
}
\examples{
bernoulli_likelihood(250, 135)
}
