% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_models.R
\name{bernoulli_posterior_hpd}
\alias{bernoulli_posterior_hpd}
\title{Exact high posterior density interval of Bernoulli model with Beta prior}
\usage{
bernoulli_posterior_hpd(n, m, alpha = 1, beta = 1)
}
\arguments{
\item{n}{Number of trials}

\item{m}{Number of successes}

\item{alpha}{First shape parameter of the Beta prior}

\item{beta}{Second shape parameter of the Beta prior}
}
\value{
A list with lower and upper bound of the HPD interval, and minimum density.
}
\description{
Exact high posterior density interval of Bernoulli model with Beta prior
}
\examples{
bernoulli_posterior_hpd(150, 120)
}
