% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_models.R
\name{bernoulli_posterior_plot}
\alias{bernoulli_posterior_plot}
\title{Plot the posterior distribution of a Bernoulli model with Beta prior}
\usage{
bernoulli_posterior_plot(n, m, alpha, beta, show_hpd = FALSE)
}
\arguments{
\item{n}{Number of trials}

\item{m}{Number of successes}

\item{alpha}{First shape parameters of the Beta prior}

\item{beta}{Second shape parameter of the Beta prior}

\item{show_hpd}{Show the HPD interval}
}
\value{

}
\description{
Plot the posterior distribution of a Bernoulli model with Beta prior
}
\examples{
bernoulli_posterior_plot(250, 139, 5, 5)
bernoulli_posterior_plot(100, 60, 2, 2)
}
