% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_models.R
\name{bernoulli_posterior_summary}
\alias{bernoulli_posterior_summary}
\title{Summary stats of posterior distribution of Bernoulli model with Beta prior}
\usage{
bernoulli_posterior_summary(n, m, alpha, beta)
}
\arguments{
\item{n}{Number of trials}

\item{m}{Number of successes}

\item{alpha}{First shape parameter of the Beta prior}

\item{beta}{Second shape parameter of the Beta prior}
}
\value{
A list with summary statistics of the posterior distribution
}
\description{
Summary stats of posterior distribution of Bernoulli model with Beta prior
}
\examples{
bernoulli_posterior_summary(3, 5)
}
